NXP-28865: merge JDBCBackend into RepositoryImpl
