--- modules/core/nuxeo-core-storage-sql/nuxeo-core-storage-sql-management/src/main/java/org/nuxeo/ecm/core/storage/sql/management/MonitoredJDBCBackend.java
@@ -1,30 +1,0 @@
-/*
- * (C) Copyright 2006-2011 Nuxeo SA (http://nuxeo.com/) and others.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * Contributors:
- *     Stephane Lacoin
- */
-
-package org.nuxeo.ecm.core.storage.sql.management;
-
-import org.nuxeo.ecm.core.storage.sql.jdbc.JDBCBackend;
-
-public class MonitoredJDBCBackend extends MonitoredBackend {
-
-    public MonitoredJDBCBackend() {
-        super(new JDBCBackend());
-    }
-
-}
