+++ modules/core/nuxeo-core-storage-sql/nuxeo-core-storage-sql/src/main/java/org/nuxeo/ecm/core/storage/sql/RepositoryDescriptor.java
@@ -182,9 +182,6 @@ public class RepositoryDescriptor {
         this.pool = pool;
     }
 
-    @XNode("backendClass")
-    public Class<? extends RepositoryBackend> backendClass;
-
     @XNode("clusterInvalidatorClass")
     public Class<? extends VCSClusterInvalidator> clusterInvalidatorClass;
 
@@ -410,7 +407,6 @@ public class RepositoryDescriptor {
         isDefault = other.isDefault;
         headless = other.headless;
         pool = other.pool == null ? null : new NuxeoConnectionManagerConfiguration(other.pool);
-        backendClass = other.backendClass;
         clusterInvalidatorClass = other.clusterInvalidatorClass;
         cachingMapperClass = other.cachingMapperClass;
         cachingMapperEnabled = other.cachingMapperEnabled;
@@ -456,9 +452,6 @@ public class RepositoryDescriptor {
             } else {
                 pool.merge(other.pool);
             }
-        }
-        if (other.backendClass != null) {
-            backendClass = other.backendClass;
         }
         if (other.clusterInvalidatorClass != null) {
             clusterInvalidatorClass = other.clusterInvalidatorClass;
